langchain Projects/1-Q&A Chatbot/requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core 
streamlit
